refactor: eliminate sideways mutations, split route adapter
Group B (Metrics as constructor args):
- Add Metrics to ServiceRegistry, pass at construction time
- Add metrics param to storage.NewService, auth.NewService,
  realtime.ManagerConfig
- Move db.SetMetrics from MetricsModule to NewServer
- Remove 4 SetMetrics sideways mutations from MetricsModule
- Remove SetMetrics methods from storage.Service, auth.Service,
  realtime.Manager

Group A (Email deferred resolution):
- Create email.LazyService for thread-safe lazy email resolution
- EmailModule registers LazyService instead of Manager
- AuthModule and TenancyModule use LazyService as email.Service
- SettingsModule creates real email.Manager with all deps,
  resolves the LazyService
- Move email.Manager creation to SettingsModule (all deps available)
- Remove email.Manager SetSettingsCache/SetSecretsService setters

Cleanup:
- Fix ClientKeyService intra-module mutation (pass cache at construction)
- Remove auth.ClientKeyService.SetSettingsCache method

Option B (Route adapter split):
- Split 977-line routes_adapter.go into 20 per-domain files
- Remaining adapter: buildAdminRouteDeps + registration boilerplate (349 lines)
