+++ internal/api/routes_adapter.go
@@ -1,636 +1,8 @@
 package api
 
 import (
-	"github.com/gofiber/fiber/v3"
-
-	"github.com/nimbleflux/fluxbase/internal/ai"
 	"github.com/nimbleflux/fluxbase/internal/api/routes"
-	"github.com/nimbleflux/fluxbase/internal/middleware"
 )
-
-func (s *Server) buildHealthRouteDeps() *routes.HealthDeps {
-	return &routes.HealthDeps{
-		Handler:      s.handleHealth,
-		OptionalAuth: s.optionalAuth,
-	}
-}
-
-func (s *Server) buildRealtimeRouteDeps() *routes.RealtimeDeps {
-	return &routes.RealtimeDeps{
-		RequireRealtimeEnabled: middleware.RequireRealtimeEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		OptionalAuth:           s.optionalAuth,
-		RequireAuth:            s.requireAuth,
-		RequireScope:           middleware.RequireScope,
-		TenantMiddleware:       s.Middleware.Tenant,
-		HandleWebSocket:        s.Realtime.Handler.HandleWebSocket,
-		HandleStats:            s.handleRealtimeStats,
-	}
-}
-
-func (s *Server) buildStorageRouteDeps() *routes.StorageDeps {
-	return &routes.StorageDeps{
-		RequireAuth:            s.requireAuth,
-		OptionalAuth:           s.optionalAuth,
-		RequireScope:           middleware.RequireScope,
-		DownloadSignedObject:   s.Storage.Handler.DownloadSignedObject,
-		GetTransformConfig:     s.Storage.Handler.GetTransformConfig,
-		ListBuckets:            s.Storage.Handler.ListBuckets,
-		CreateBucket:           s.Storage.Handler.CreateBucket,
-		UpdateBucketSettings:   s.Storage.Handler.UpdateBucketSettings,
-		DeleteBucket:           s.Storage.Handler.DeleteBucket,
-		ListFiles:              s.Storage.Handler.ListFiles,
-		MultipartUpload:        s.Storage.Handler.MultipartUpload,
-		ShareObject:            s.Storage.Handler.ShareObject,
-		RevokeShare:            s.Storage.Handler.RevokeShare,
-		ListShares:             s.Storage.Handler.ListShares,
-		GenerateSignedURL:      s.Storage.Handler.GenerateSignedURL,
-		StreamUpload:           s.Storage.Handler.StreamUpload,
-		StorageUploadLimiter:   middleware.StorageUploadLimiter(s.sharedMiddlewareStorage),
-		InitChunkedUpload:      s.Storage.Handler.InitChunkedUpload,
-		UploadChunk:            s.Storage.Handler.UploadChunk,
-		CompleteChunkedUpload:  s.Storage.Handler.CompleteChunkedUpload,
-		GetChunkedUploadStatus: s.Storage.Handler.GetChunkedUploadStatus,
-		AbortChunkedUpload:     s.Storage.Handler.AbortChunkedUpload,
-		UploadFile:             s.Storage.Handler.UploadFile,
-		DownloadFile:           s.Storage.Handler.DownloadFile,
-		DeleteFile:             s.Storage.Handler.DeleteFile,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildRESTRouteDeps() *routes.RESTDeps {
-	return &routes.RESTDeps{
-		RequireAuth:        s.requireAuth,
-		RequireScope:       middleware.RequireScope,
-		HandleTables:       s.rest.HandleDynamicTable,
-		HandleQuery:        s.rest.HandleDynamicQuery,
-		HandleById:         s.rest.HandleDynamicTableById,
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildGraphQLRouteDeps() *routes.GraphQLDeps {
-	if s.GraphQL.Handler == nil {
-		return nil
-	}
-	return &routes.GraphQLDeps{
-		OptionalAuth:     s.optionalAuth,
-		HandleGraphQL:    s.GraphQL.Handler.HandleGraphQL,
-		HandleIntrospect: s.GraphQL.Handler.HandleIntrospection,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildVectorRouteDeps() *routes.VectorDeps {
-	if s.AI.VectorHandler == nil {
-		return nil
-	}
-	return &routes.VectorDeps{
-		RequireAuth:        s.requireAuth,
-		TenantMiddleware:   s.Middleware.Tenant,
-		HandleCapabilities: s.AI.VectorHandler.HandleGetCapabilities,
-		HandleEmbed:        s.AI.VectorHandler.HandleEmbed,
-		HandleSearch:       s.AI.VectorHandler.HandleSearch,
-	}
-}
-
-func (s *Server) buildRPCRouteDeps() *routes.RPCDeps {
-	if s.RPC.Handler == nil {
-		return nil
-	}
-	return &routes.RPCDeps{
-		RequireRPCEnabled: middleware.RequireRPCEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		OptionalAuth:      s.optionalAuth,
-		RequireScope:      middleware.RequireScope,
-		ListProcedures:    s.RPC.Handler.ListPublicProcedures,
-		Invoke:            s.RPC.Handler.Invoke,
-		GetExecution:      s.RPC.Handler.GetPublicExecution,
-		GetExecutionLogs:  s.RPC.Handler.GetPublicExecutionLogs,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildAIRouteDeps() *routes.AIDeps {
-	if s.AI.Chat == nil || s.AI.Handler == nil {
-		return nil
-	}
-	return &routes.AIDeps{
-		RequireAIEnabled:       middleware.RequireAIEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		OptionalAuth:           s.optionalAuth,
-		RequireAuth:            s.requireAuth,
-		TenantMiddleware:       s.Middleware.Tenant,
-		HandleWebSocket:        s.AI.Chat.HandleWebSocket,
-		ListPublicChatbots:     s.AI.Handler.ListPublicChatbots,
-		LookupChatbotByName:    s.AI.Handler.LookupChatbotByName,
-		GetPublicChatbot:       s.AI.Handler.GetPublicChatbot,
-		ListUserConversations:  s.AI.Handler.ListUserConversations,
-		GetUserConversation:    s.AI.Handler.GetUserConversation,
-		DeleteUserConversation: s.AI.Handler.DeleteUserConversation,
-		UpdateUserConversation: s.AI.Handler.UpdateUserConversation,
-	}
-}
-
-func (s *Server) buildSettingsRouteDeps() *routes.SettingsDeps {
-	return &routes.SettingsDeps{
-		OptionalAuth: s.optionalAuth,
-		GetSetting:   s.Settings.Handler.GetSetting,
-		GetSettings:  s.Settings.Handler.GetSettings,
-		BatchGet:     s.Settings.Handler.GetSettings,
-	}
-}
-
-func (s *Server) buildUserSettingsRouteDeps() *routes.UserSettingsDeps {
-	return &routes.UserSettingsDeps{
-		RequireAuth:       s.requireAuth,
-		TenantMiddleware:  s.Middleware.Tenant,
-		ListSettings:      s.Settings.User.ListSettings,
-		GetUserOwnSetting: s.Settings.User.GetUserOwnSetting,
-		GetSystemSetting:  s.Settings.User.GetSystemSettingPublic,
-		GetSetting:        s.Settings.User.GetSetting,
-		SetSetting:        s.Settings.User.SetSetting,
-		DeleteSetting:     s.Settings.User.DeleteSetting,
-		CreateSecret:      s.Settings.User.CreateSecret,
-		ListSecrets:       s.Settings.User.ListSecrets,
-		GetSecret:         s.Settings.User.GetSecret,
-		UpdateSecret:      s.Settings.User.UpdateSecret,
-		DeleteSecret:      s.Settings.User.DeleteSecret,
-	}
-}
-
-func (s *Server) buildDashboardAuthRouteDeps() *routes.DashboardAuthDeps {
-	return &routes.DashboardAuthDeps{
-		SetupLimiter:    middleware.AdminSetupLimiterWithConfig(s.config.Security.AdminSetupRateLimit, s.config.Security.AdminSetupRateWindow, s.sharedMiddlewareStorage),
-		LoginLimiter:    middleware.AdminLoginLimiterWithConfig(s.config.Security.AdminLoginRateLimit, s.config.Security.AdminLoginRateWindow, s.sharedMiddlewareStorage),
-		GetSetupStatus:  s.Auth.AdminHandler.GetSetupStatus,
-		InitialSetup:    s.Auth.AdminHandler.InitialSetup,
-		AdminLogin:      s.Auth.AdminHandler.AdminLogin,
-		RefreshToken:    s.Auth.AdminHandler.AdminRefreshToken,
-		UnifiedAuth:     UnifiedAuthMiddleware(s.Auth.Handler.authService, s.Auth.DashboardHandler.jwtManager, s.db),
-		AdminLogout:     s.Auth.AdminHandler.AdminLogout,
-		GetCurrentAdmin: s.Auth.AdminHandler.GetCurrentAdmin,
-	}
-}
-
-func (s *Server) buildOpenAPIRouteDeps() *routes.OpenAPIDeps {
-	return &routes.OpenAPIDeps{
-		OptionalAuth:    s.optionalAuth,
-		TenantContext:   s.Middleware.Tenant,
-		TenantDBContext: s.Middleware.TenantDB,
-		GetOpenAPISpec:  NewOpenAPIHandler(s.db).GetOpenAPISpec,
-	}
-}
-
-func (s *Server) buildAuthRouteDeps() *routes.AuthDeps {
-	rateLimiters := map[string]fiber.Handler{
-		"signup":         middleware.AuthSignupLimiterWithConfig(s.config.Security.AuthSignupRateLimit, s.config.Security.AuthSignupRateWindow, s.sharedMiddlewareStorage),
-		"login":          middleware.AuthLoginLimiterWithConfig(s.config.Security.AuthLoginRateLimit, s.config.Security.AuthLoginRateWindow, s.sharedMiddlewareStorage),
-		"refresh":        middleware.AuthRefreshLimiterWithConfig(s.config.Security.AuthRefreshRateLimit, s.config.Security.AuthRefreshRateWindow, s.sharedMiddlewareStorage),
-		"magiclink":      middleware.AuthMagicLinkLimiterWithConfig(s.config.Security.AuthMagicLinkRateLimit, s.config.Security.AuthMagicLinkRateWindow, s.sharedMiddlewareStorage),
-		"password_reset": middleware.AuthPasswordResetLimiterWithConfig(s.config.Security.AuthPasswordResetRateLimit, s.config.Security.AuthPasswordResetRateWindow, s.sharedMiddlewareStorage),
-		"otp":            middleware.AuthMagicLinkLimiterWithConfig(s.config.Security.AuthMagicLinkRateLimit, s.config.Security.AuthMagicLinkRateWindow, s.sharedMiddlewareStorage),
-		"2fa":            middleware.Auth2FALimiterWithConfig(s.config.Security.Auth2FARateLimit, s.config.Security.Auth2FARateWindow, s.sharedMiddlewareStorage),
-	}
-
-	return &routes.AuthDeps{
-		AuthMiddleware:            AuthMiddleware(s.Auth.Handler.authService),
-		TenantMiddleware:          s.Middleware.Tenant,
-		RequireRole:               RequireRole,
-		RequireScope:              middleware.RequireScope,
-		RateLimiters:              rateLimiters,
-		GetCSRFToken:              s.Auth.Handler.GetCSRFToken,
-		GetCaptchaConfig:          s.Auth.Handler.GetCaptchaConfig,
-		CheckCaptcha:              s.Auth.Handler.CheckCaptcha,
-		GetAuthConfig:             s.Auth.Handler.GetAuthConfig,
-		SignUp:                    s.Auth.Handler.SignUp,
-		SignIn:                    s.Auth.Handler.SignIn,
-		RefreshToken:              s.Auth.Handler.RefreshToken,
-		SendMagicLink:             s.Auth.Handler.SendMagicLink,
-		VerifyMagicLink:           s.Auth.Handler.VerifyMagicLink,
-		RequestPasswordReset:      s.Auth.Handler.RequestPasswordReset,
-		ResetPassword:             s.Auth.Handler.ResetPassword,
-		VerifyPasswordReset:       s.Auth.Handler.VerifyPasswordResetToken,
-		VerifyEmail:               s.Auth.Handler.VerifyEmail,
-		ResendVerification:        s.Auth.Handler.ResendVerificationEmail,
-		VerifyTOTP:                s.Auth.Handler.VerifyTOTP,
-		SendOTP:                   s.Auth.Handler.SendOTP,
-		VerifyOTP:                 s.Auth.Handler.VerifyOTP,
-		ResendOTP:                 s.Auth.Handler.ResendOTP,
-		SignInWithIDToken:         s.Auth.Handler.SignInWithIDToken,
-		SignOut:                   s.Auth.Handler.SignOut,
-		GetUser:                   s.Auth.Handler.GetUser,
-		UpdateUser:                s.Auth.Handler.UpdateUser,
-		StartImpersonation:        s.Auth.Handler.StartImpersonation,
-		StartAnonImpersonation:    s.Auth.Handler.StartAnonImpersonation,
-		StopImpersonation:         s.Auth.Handler.StopImpersonation,
-		GetActiveImpersonation:    s.Auth.Handler.GetActiveImpersonation,
-		ListImpersonationSessions: s.Auth.Handler.ListImpersonationSessions,
-		SetupTOTP:                 s.Auth.Handler.SetupTOTP,
-		EnableTOTP:                s.Auth.Handler.EnableTOTP,
-		DisableTOTP:               s.Auth.Handler.DisableTOTP,
-		GetTOTPStatus:             s.Auth.Handler.GetTOTPStatus,
-		GetUserIdentities:         s.Auth.Handler.GetUserIdentities,
-		LinkIdentity:              s.Auth.Handler.LinkIdentity,
-		UnlinkIdentity:            s.Auth.Handler.UnlinkIdentity,
-		Reauthenticate:            s.Auth.Handler.Reauthenticate,
-		ListOAuthProviders:        s.Auth.OAuth.ListEnabledProviders,
-		OAuthAuthorize:            s.Auth.OAuth.Authorize,
-		OAuthCallback:             s.Auth.OAuth.Callback,
-		GetSPMetadata:             s.Auth.SAMLProvider.GetSPMetadata,
-		StartServiceImpersonation: s.Auth.Handler.StartServiceImpersonation,
-		GetProviderToken:          s.Auth.OAuth.GetProviderToken,
-		OAuthLogout:               s.Auth.OAuth.Logout,
-		OAuthLogoutCallback:       s.Auth.OAuth.LogoutCallback,
-		ListSAMLProviders:         s.Auth.SAML.ListSAMLProviders,
-		InitiateSAMLLogin:         s.Auth.SAML.InitiateSAMLLogin,
-		HandleSAMLAssertion:       s.Auth.SAML.HandleSAMLAssertion,
-		HandleSAMLLogout:          s.Auth.SAML.HandleSAMLLogout,
-		InitiateSAMLLogout:        s.Auth.SAML.InitiateSAMLLogout,
-	}
-}
-
-func (s *Server) buildInternalAIRouteDeps() *routes.InternalAIDeps {
-	if s.AI.Internal == nil {
-		return nil
-	}
-	return &routes.InternalAIDeps{
-		RequireInternal:     middleware.RequireInternal(),
-		RequireAuth:         s.requireAuth,
-		HandleChat:          s.AI.Internal.HandleChat,
-		HandleEmbed:         s.AI.Internal.HandleEmbed,
-		HandleListProviders: s.AI.Internal.HandleListProviders,
-	}
-}
-
-func (s *Server) buildGitHubWebhookRouteDeps() *routes.GitHubWebhookDeps {
-	if s.Branching.GitHub == nil {
-		return nil
-	}
-	return &routes.GitHubWebhookDeps{
-		GitHubWebhookLimiter: middleware.GitHubWebhookLimiter(s.sharedMiddlewareStorage),
-		HandleWebhook:        s.Branching.GitHub.HandleWebhook,
-	}
-}
-
-func (s *Server) buildInvitationRouteDeps() *routes.InvitationDeps {
-	return &routes.InvitationDeps{
-		ValidateInvitation: s.Auth.Invitation.ValidateInvitation,
-		AcceptInvitation:   s.Auth.Invitation.AcceptInvitation,
-	}
-}
-
-func (s *Server) buildWebhookRouteDeps() *routes.WebhookDeps {
-	return &routes.WebhookDeps{
-		RequireAuth:    s.requireAuth,
-		RequireScope:   middleware.RequireScope,
-		ListWebhooks:   s.Webhook.Handler.ListWebhooks,
-		GetWebhook:     s.Webhook.Handler.GetWebhook,
-		ListDeliveries: s.Webhook.Handler.ListDeliveries,
-		CreateWebhook:  s.Webhook.Handler.CreateWebhook,
-		UpdateWebhook:  s.Webhook.Handler.UpdateWebhook,
-		DeleteWebhook:  s.Webhook.Handler.DeleteWebhook,
-		TestWebhook:    s.Webhook.Handler.TestWebhook,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildMonitoringRouteDeps() *routes.MonitoringDeps {
-	return &routes.MonitoringDeps{
-		RequireAuth:        s.requireAuth,
-		RequireScope:       middleware.RequireScope,
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-		GetMetrics:         s.Monitoring.Handler.GetMetrics,
-		GetHealth:          s.Monitoring.Handler.GetHealth,
-		GetLogs:            s.Monitoring.Handler.GetLogs,
-	}
-}
-
-func (s *Server) buildFunctionsRouteDeps() *routes.FunctionsDeps {
-	if s.Functions.Handler == nil {
-		return nil
-	}
-	return &routes.FunctionsDeps{
-		RequireFunctionsEnabled: middleware.RequireFunctionsEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		RequireAuth:             s.requireAuth,
-		OptionalAuth:            s.optionalAuth,
-		RequireScope:            middleware.RequireScope,
-		TenantMiddleware:        s.Middleware.Tenant,
-		ListFunctions:           s.Functions.Handler.ListFunctions,
-		GetFunction:             s.Functions.Handler.GetFunction,
-		CreateFunction:          s.Functions.Handler.CreateFunction,
-		UpdateFunction:          s.Functions.Handler.UpdateFunction,
-		DeleteFunction:          s.Functions.Handler.DeleteFunction,
-		InvokeFunction:          s.Functions.Handler.InvokeFunction,
-		GetExecutions:           s.Functions.Handler.GetExecutions,
-		ListSharedModules:       s.Functions.Handler.ListSharedModules,
-		GetSharedModule:         s.Functions.Handler.GetSharedModule,
-		CreateSharedModule:      s.Functions.Handler.CreateSharedModule,
-		UpdateSharedModule:      s.Functions.Handler.UpdateSharedModule,
-		DeleteSharedModule:      s.Functions.Handler.DeleteSharedModule,
-	}
-}
-
-func (s *Server) buildJobsRouteDeps() *routes.JobsDeps {
-	if s.Jobs.Handler == nil {
-		return nil
-	}
-	return &routes.JobsDeps{
-		RequireJobsEnabled: middleware.RequireJobsEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		RequireAuth:        s.requireAuth,
-		SubmitJob:          s.Jobs.Handler.SubmitJob,
-		GetJob:             s.Jobs.Handler.GetJob,
-		ListJobs:           s.Jobs.Handler.ListJobs,
-		CancelJob:          s.Jobs.Handler.CancelJob,
-		RetryJob:           s.Jobs.Handler.RetryJob,
-		GetJobLogsUser:     s.Jobs.Handler.GetJobLogsUser,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildBranchRouteDeps() *routes.BranchDeps {
-	if s.Branching.Handler == nil || !s.config.Branching.Enabled {
-		return nil
-	}
-
-	return &routes.BranchDeps{
-		GetActiveBranch:    s.Branching.Handler.GetActiveBranch,
-		SetActiveBranch:    s.Branching.Handler.SetActiveBranch,
-		ResetActiveBranch:  s.Branching.Handler.ResetActiveBranch,
-		GetPoolStats:       s.Branching.Handler.GetPoolStats,
-		CreateBranch:       s.Branching.Handler.CreateBranch,
-		ListBranches:       s.Branching.Handler.ListBranches,
-		GetBranch:          s.Branching.Handler.GetBranch,
-		DeleteBranch:       s.Branching.Handler.DeleteBranch,
-		ResetBranch:        s.Branching.Handler.ResetBranch,
-		GetBranchActivity:  s.Branching.Handler.GetBranchActivity,
-		ListBranchAccess:   s.Branching.Handler.ListBranchAccess,
-		GrantBranchAccess:  s.Branching.Handler.GrantBranchAccess,
-		RevokeBranchAccess: s.Branching.Handler.RevokeBranchAccess,
-		ListGitHubConfigs:  s.Branching.Handler.ListGitHubConfigs,
-		UpsertGitHubConfig: s.Branching.Handler.UpsertGitHubConfig,
-		DeleteGitHubConfig: s.Branching.Handler.DeleteGitHubConfig,
-	}
-}
-
-func (s *Server) buildClientKeysRouteDeps() *routes.ClientKeysDeps {
-	return &routes.ClientKeysDeps{
-		RequireAuth:                      s.requireAuth,
-		RequireAdminIfClientKeysDisabled: middleware.RequireAdminIfClientKeysDisabled(s.Auth.Handler.authService.GetSettingsCache()),
-		RequireScope:                     middleware.RequireScope,
-		TenantMiddleware:                 s.Middleware.Tenant,
-		ListClientKeys:                   s.Auth.ClientKeyHandler.ListClientKeys,
-		GetClientKey:                     s.Auth.ClientKeyHandler.GetClientKey,
-		CreateClientKey:                  s.Auth.ClientKeyHandler.CreateClientKey,
-		UpdateClientKey:                  s.Auth.ClientKeyHandler.UpdateClientKey,
-		DeleteClientKey:                  s.Auth.ClientKeyHandler.DeleteClientKey,
-		RevokeClientKey:                  s.Auth.ClientKeyHandler.RevokeClientKey,
-	}
-}
-
-func (s *Server) buildSecretsRouteDeps() *routes.SecretsDeps {
-	if s.Secrets.Handler == nil {
-		return nil
-	}
-	return &routes.SecretsDeps{
-		RequireAuth:        s.requireAuth,
-		RequireScope:       middleware.RequireScope,
-		ListSecrets:        s.Secrets.Handler.ListSecrets,
-		GetStats:           s.Secrets.Handler.GetStats,
-		GetSecretByName:    s.Secrets.Handler.GetSecretByName,
-		GetVersionsByName:  s.Secrets.Handler.GetVersionsByName,
-		UpdateSecretByName: s.Secrets.Handler.UpdateSecretByName,
-		DeleteSecretByName: s.Secrets.Handler.DeleteSecretByName,
-		RollbackByName:     s.Secrets.Handler.RollbackByName,
-		GetSecret:          s.Secrets.Handler.GetSecret,
-		GetVersions:        s.Secrets.Handler.GetVersions,
-		CreateSecret:       s.Secrets.Handler.CreateSecret,
-		UpdateSecret:       s.Secrets.Handler.UpdateSecret,
-		DeleteSecret:       s.Secrets.Handler.DeleteSecret,
-		RollbackToVersion:  s.Secrets.Handler.RollbackToVersion,
-
-		TenantMiddleware:   s.Middleware.Tenant,
-		TenantDBMiddleware: s.Middleware.TenantDB,
-	}
-}
-
-func (s *Server) buildSyncRouteDeps() *routes.SyncDeps {
-	deps := &routes.SyncDeps{
-		RequireSyncAuth: UnifiedAuthMiddleware(s.Auth.Handler.authService, s.Auth.DashboardHandler.jwtManager, s.db),
-		RequireRole:     RequireRole("admin", "instance_admin", "service_role"),
-		TenantContext:   s.Middleware.Tenant,
-	}
-
-	// Functions sync
-	if s.Functions.Handler != nil {
-		deps.RequireFunctionsSyncIPAllowlist = middleware.RequireSyncIPAllowlist(s.config.Functions.SyncAllowedIPRanges, "functions", &s.config.Server)
-		deps.SyncFunctions = s.Functions.Handler.SyncFunctions
-	}
-
-	// Jobs sync
-	if s.Jobs.Handler != nil {
-		deps.RequireJobsSyncIPAllowlist = middleware.RequireSyncIPAllowlist(s.config.Jobs.SyncAllowedIPRanges, "jobs", &s.config.Server)
-		deps.SyncJobs = s.Jobs.Handler.SyncJobs
-	}
-
-	// AI sync
-	if s.AI.Handler != nil {
-		deps.RequireAIEnabled = middleware.RequireAIEnabled(s.Auth.Handler.authService.GetSettingsCache())
-		deps.RequireAISyncIPAllowlist = middleware.RequireSyncIPAllowlist(s.config.AI.SyncAllowedIPRanges, "ai", &s.config.Server)
-		deps.SyncChatbots = s.AI.Handler.SyncChatbots
-	}
-
-	// RPC sync
-	if s.RPC.Handler != nil {
-		deps.RequireRPCEnabled = middleware.RequireRPCEnabled(s.Auth.Handler.authService.GetSettingsCache())
-		deps.RequireRPCSyncIPAllowlist = middleware.RequireSyncIPAllowlist(s.config.RPC.SyncAllowedIPRanges, "rpc", &s.config.Server)
-		deps.SyncProcedures = s.RPC.Handler.SyncProcedures
-	}
-
-	return deps
-}
-
-func (s *Server) buildDashboardUserAuthRouteDeps() *routes.DashboardUserAuthDeps {
-	return &routes.DashboardUserAuthDeps{
-		RequireDashboardAuth:     s.Auth.DashboardHandler.RequireDashboardAuth,
-		TenantMiddleware:         s.Middleware.Tenant,
-		Signup:                   s.Auth.DashboardHandler.Signup,
-		Login:                    s.Auth.DashboardHandler.Login,
-		RefreshToken:             s.Auth.DashboardHandler.RefreshToken,
-		VerifyTOTP:               s.Auth.DashboardHandler.VerifyTOTP,
-		RequestPasswordReset:     s.Auth.DashboardHandler.RequestPasswordReset,
-		VerifyPasswordResetToken: s.Auth.DashboardHandler.VerifyPasswordResetToken,
-		ConfirmPasswordReset:     s.Auth.DashboardHandler.ConfirmPasswordReset,
-		GetSSOProviders:          s.Auth.DashboardHandler.GetSSOProviders,
-		InitiateOAuthLogin:       s.Auth.DashboardHandler.InitiateOAuthLogin,
-		OAuthCallback:            s.Auth.DashboardHandler.OAuthCallback,
-		InitiateSAMLLogin:        s.Auth.DashboardHandler.InitiateSAMLLogin,
-		SAMLACSCallback:          s.Auth.DashboardHandler.SAMLACSCallback,
-		GetCurrentUser:           s.Auth.DashboardHandler.GetCurrentUser,
-		UpdateProfile:            s.Auth.DashboardHandler.UpdateProfile,
-		ChangePassword:           s.Auth.DashboardHandler.ChangePassword,
-		DeleteAccount:            s.Auth.DashboardHandler.DeleteAccount,
-		SetupTOTP:                s.Auth.DashboardHandler.SetupTOTP,
-		EnableTOTP:               s.Auth.DashboardHandler.EnableTOTP,
-		DisableTOTP:              s.Auth.DashboardHandler.DisableTOTP,
-	}
-}
-
-func (s *Server) buildCustomMCPRouteDeps() *routes.CustomMCPDeps {
-	if s.MCP.CustomHandler == nil {
-		return nil
-	}
-	return &routes.CustomMCPDeps{
-		RequireAuth:      s.requireAuth,
-		RequireAdmin:     middleware.RequireAdmin(),
-		TenantMiddleware: s.Middleware.Tenant,
-		GetConfig:        s.MCP.CustomHandler.GetConfig,
-		ListTools:        s.MCP.CustomHandler.ListTools,
-		CreateTool:       s.MCP.CustomHandler.CreateTool,
-		SyncTool:         s.MCP.CustomHandler.SyncTool,
-		GetTool:          s.MCP.CustomHandler.GetTool,
-		UpdateTool:       s.MCP.CustomHandler.UpdateTool,
-		DeleteTool:       s.MCP.CustomHandler.DeleteTool,
-		TestTool:         s.MCP.CustomHandler.TestTool,
-		ListResources:    s.MCP.CustomHandler.ListResources,
-		CreateResource:   s.MCP.CustomHandler.CreateResource,
-		SyncResource:     s.MCP.CustomHandler.SyncResource,
-		GetResource:      s.MCP.CustomHandler.GetResource,
-		UpdateResource:   s.MCP.CustomHandler.UpdateResource,
-		DeleteResource:   s.MCP.CustomHandler.DeleteResource,
-		TestResource:     s.MCP.CustomHandler.TestResource,
-	}
-}
-
-func (s *Server) buildMCPRouteDeps() *routes.MCPDeps {
-	if s.MCP.Handler == nil {
-		return nil
-	}
-	return &routes.MCPDeps{
-		BasePath:         s.config.MCP.BasePath,
-		MCPAuth:          s.createMCPAuthMiddleware(),
-		TenantMiddleware: s.Middleware.Tenant,
-		HandlePost:       s.MCP.Handler.HandlePost,
-		HandleGet:        s.MCP.Handler.HandleGet,
-		HandleHealth:     s.MCP.Handler.HandleHealth,
-	}
-}
-
-func (s *Server) buildMCPOAuthRouteDeps() *routes.MCPOAuthDeps {
-	if s.MCP.OAuth == nil {
-		return nil
-	}
-	return &routes.MCPOAuthDeps{
-		BasePath:                          s.config.MCP.BasePath,
-		HandleAuthorizationServerMetadata: s.MCP.OAuth.HandleAuthorizationServerMetadata,
-		HandleProtectedResourceMetadata:   s.MCP.OAuth.HandleProtectedResourceMetadata,
-		HandleClientRegistration:          s.MCP.OAuth.HandleClientRegistration,
-		HandleAuthorize:                   s.MCP.OAuth.HandleAuthorize,
-		HandleAuthorizeConsent:            s.MCP.OAuth.HandleAuthorizeConsent,
-		HandleToken:                       s.MCP.OAuth.HandleToken,
-		HandleRevoke:                      s.MCP.OAuth.HandleRevoke,
-	}
-}
-
-func (s *Server) buildMigrationsRouteDeps() *routes.MigrationsDeps {
-	if s.Schema.Migrations == nil || !s.config.Migrations.Enabled {
-		return nil
-	}
-
-	var tenantPoolProvider middleware.MigrationsTenantPoolProvider
-	if s.Tenancy.Manager != nil && s.Tenancy.Manager.GetRouter() != nil {
-		tenantPoolProvider = s.Tenancy.Manager.GetRouter()
-	}
-
-	return &routes.MigrationsDeps{
-		SecurityMiddleware: middleware.RequireMigrationsFullSecurityWithTenantProvider(
-			&s.config.Migrations,
-			&s.config.Server,
-			s.db.Pool(),
-			s.Auth.Handler.authService,
-			s.config.Security.ServiceRoleRateLimit,
-			s.config.Security.ServiceRoleRateWindow,
-			s.sharedMiddlewareStorage,
-			tenantPoolProvider,
-		),
-		TenantContext:     s.Middleware.Tenant,
-		RequireRole:       RequireRole,
-		CreateMigration:   s.Schema.Migrations.CreateMigration,
-		ListMigrations:    s.Schema.Migrations.ListMigrations,
-		GetMigration:      s.Schema.Migrations.GetMigration,
-		UpdateMigration:   s.Schema.Migrations.UpdateMigration,
-		DeleteMigration:   s.Schema.Migrations.DeleteMigration,
-		ApplyMigration:    s.Schema.Migrations.ApplyMigration,
-		RollbackMigration: s.Schema.Migrations.RollbackMigration,
-		ApplyPending:      s.Schema.Migrations.ApplyPending,
-		SyncMigrations:    s.Schema.Migrations.SyncMigrations,
-		GetExecutions:     s.Schema.Migrations.GetExecutions,
-	}
-}
-
-// knowledgeBaseDisabledHandler returns a handler that responds with "AI not enabled" error
-func knowledgeBaseDisabledHandler(c fiber.Ctx) error {
-	return SendFeatureDisabled(c, "AI features")
-}
-
-func (s *Server) buildKnowledgeBaseRouteDeps() *routes.KnowledgeBaseDeps {
-	deps := &routes.KnowledgeBaseDeps{
-		RequireAIEnabled: middleware.RequireAIEnabled(s.Auth.Handler.authService.GetSettingsCache()),
-		RequireAuth:      s.requireAuth,
-		TenantMiddleware: s.Middleware.Tenant,
-	}
-
-	// If AI/knowledge base storage is not available, use stub handlers
-	// This ensures routes return a proper error instead of 404
-	if s.AI.KBStorage == nil {
-		deps.ListKBs = knowledgeBaseDisabledHandler
-		deps.CreateKB = knowledgeBaseDisabledHandler
-		deps.GetKB = knowledgeBaseDisabledHandler
-		deps.ShareKB = knowledgeBaseDisabledHandler
-		deps.ListPermissions = knowledgeBaseDisabledHandler
-		deps.RevokePermission = knowledgeBaseDisabledHandler
-		deps.ListDocuments = knowledgeBaseDisabledHandler
-		deps.GetDocument = knowledgeBaseDisabledHandler
-		deps.AddDocument = knowledgeBaseDisabledHandler
-		deps.UploadDocument = knowledgeBaseDisabledHandler
-		deps.DeleteDocument = knowledgeBaseDisabledHandler
-		deps.SearchKB = knowledgeBaseDisabledHandler
-		return deps
-	}
-
-	handler := ai.NewUserKnowledgeBaseHandler(s.AI.KBStorage)
-	if s.AI.DocProcessor != nil {
-		handler = ai.NewUserKnowledgeBaseHandlerWithProcessor(s.AI.KBStorage, s.AI.DocProcessor)
-	}
-
-	deps.ListKBs = handler.ListMyKnowledgeBases
-	deps.CreateKB = handler.CreateMyKnowledgeBase
-	deps.GetKB = handler.GetMyKnowledgeBase
-	deps.ShareKB = handler.ShareKnowledgeBase
-	deps.ListPermissions = handler.ListPermissions
-	deps.RevokePermission = handler.RevokePermission
-
-	if s.AI.DocProcessor != nil {
-		deps.ListDocuments = handler.ListMyDocuments
-		deps.GetDocument = handler.GetMyDocument
-		deps.AddDocument = handler.AddMyDocument
-		deps.UploadDocument = handler.UploadMyDocument
-		deps.DeleteDocument = handler.DeleteMyDocument
-		deps.SearchKB = handler.SearchMyKB
-	}
-
-	return deps
-}
 
 func (s *Server) buildAdminRouteDeps() *routes.AdminDeps {
 	unifiedAuth := UnifiedAuthMiddleware(s.Auth.Handler.authService, s.Auth.DashboardHandler.jwtManager, s.db)
