+++ internal/api/routes_health.go
@@ -1,0 +1,12 @@
+package api
+
+import (
+	"github.com/nimbleflux/fluxbase/internal/api/routes"
+)
+
+func (s *Server) buildHealthRouteDeps() *routes.HealthDeps {
+	return &routes.HealthDeps{
+		Handler:      s.handleHealth,
+		OptionalAuth: s.optionalAuth,
+	}
+}
